pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Howard Halim", email = "howardhalim@gmail.com"},
]
dependencies = [
    "matplotlib>=3.7.0",
    "clip-by-openai>=1.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}
